requirements.txt
begins
colorama
requests
schedule
